SEAR-3 refactoring: got rid of ignore=true in mappers, started decomposing of Utils. java
