+++ build.gradle
@@ -339,7 +339,8 @@ task publishDocker(dependsOn: [pushDockerLatest, pushDockerVersionTagged]) {
 
 tasks.withType(JavaCompile) {
     options.compilerArgs = [
-            '-Amapstruct.suppressGeneratorTimestamp=true'
+            '-Amapstruct.suppressGeneratorTimestamp=true',
+            '-Amapstruct.unmappedTargetPolicy=IGNORE'
     ]
 }
