+++ src/main/java/gov/ca/cwds/cals/service/mapper/FasFacilityMapper.java
@@ -19,25 +19,7 @@ import org.mapstruct.factory.Mappers;
 public interface FasFacilityMapper {
   FasFacilityMapper INSTANCE = Mappers.getMapper(FasFacilityMapper.class);
 
-  @Mapping(target = "messages", ignore = true)
-  @Mapping(target = "phone", ignore = true)
-  @Mapping(target = "visits", ignore = true)
-  @Mapping(target = "address", ignore = true)
-  @Mapping(target = "county", ignore = true)
-  @Mapping(target = "emailAddress", ignore = true)
   @Mapping(target = "originalApplicationRecievedDate", ignore = true)
-  @Mapping(target = "licenseEffectiveDate", ignore = true)
-  @Mapping(target = "capacity", ignore = true)
-  @Mapping(target = "licenseNumber", ignore = true)
-  @Mapping(target = "status", ignore = true)
-  @Mapping(target = "districtOffice", ignore = true)
-  @Mapping(target = "assignedWorker", ignore = true)
-  @Mapping(target = "licenseeType", ignore = true)
-  @Mapping(target = "licenseeName", ignore = true)
-  @Mapping(target = "name", ignore = true)
-  @Mapping(target = "type", ignore = true)
-  @Mapping(target = "id", ignore = true)
-  @Mapping(target = "href", ignore = true)
   @Mapping(target = "lastVisitDate", source = "facLastVisitDate")
   @Mapping(target = "lastVisitReason", source = "facilityLastVisitReason", qualifiedByName = "facilityVisitReason")
   @Mapping(target = "lastDeferredVisitDate", source = "facLastDeferVisitDate")
