+++ src/main/java/gov/ca/cwds/cals/service/mapper/PersonMapper.java
@@ -1,8 +1,8 @@
 package gov.ca.cwds.cals.service.mapper;
 
 import gov.ca.cwds.cals.service.dto.PersonDTO;
-import gov.ca.cwds.data.legacy.cms.entity.Client;
 import gov.ca.cwds.data.legacy.cms.entity.BaseStaffPerson;
+import gov.ca.cwds.data.legacy.cms.entity.Client;
 import java.time.LocalDate;
 import java.time.Period;
 import java.util.List;
@@ -24,21 +24,10 @@ public interface PersonMapper {
     @Mapping(target = "lastName", source = "commonLastName")
     @Mapping(target = "gender", source = "gender.code")
     @Mapping(target = "dateOfBirth", source = "birthDate")
-
-    @Mapping(target = "id",          ignore = true)
-    @Mapping(target = "age",         ignore = true)
-    @Mapping(target = "ssn",         ignore = true)
-    @Mapping(target = "ethnicityId", ignore = true)
     PersonDTO toPerson(Client client);
 
     @Mapping(target = "firstName", source = "firstName")
     @Mapping(target = "lastName", source = "lastName")
-    @Mapping(target = "id",          ignore = true)
-    @Mapping(target = "gender",      ignore = true)
-    @Mapping(target = "age",         ignore = true)
-    @Mapping(target = "dateOfBirth", ignore = true)
-    @Mapping(target = "ssn",         ignore = true)
-    @Mapping(target = "ethnicityId", ignore = true)
     PersonDTO toPersonDTO(BaseStaffPerson staffPerson);
 
     @AfterMapping
