+++ src/main/java/gov/ca/cwds/cals/service/mapper/PhoneMapper.java
@@ -18,21 +18,16 @@ import org.mapstruct.Mapping;
 @DecoratedWith(PhoneMapperDecorator.class)
 public interface PhoneMapper {
 
-    @Mapping(target = "id", ignore = true)
     @Mapping(target = "relation", constant = PRIMARY)
     @Mapping(target = "type", constant = "Cell")
     @Mapping(target = "number", source = "facPhoneNbr")
     PersonPhoneDTO lisFacilityToPhoneDTO(LisFacFile lisFacFile);
 
-    @Mapping(target = "id", ignore = true)
     @Mapping(target = "relation", constant = PRIMARY)
-    @Mapping(target = "type", ignore = true)
     @Mapping(target = "number", source = "prmTelNo")
     PersonPhoneDTO toPrimaryPhoneDTO(BasePlacementHome placementHome);
 
-    @Mapping(target = "id", ignore = true)
     @Mapping(target = "relation", constant = ALTERNATE)
-    @Mapping(target = "type", ignore = true)
     @Mapping(target = "number", source = "bckTelNo")
     PersonPhoneDTO toAlternatePhoneDTO(BasePlacementHome placementHome);
