+++ src/main/java/gov/ca/cwds/cals/service/mapper/PlacementHomeMapper.java
@@ -3,8 +3,12 @@ package gov.ca.cwds.cals.service.mapper;
 import gov.ca.cwds.cals.service.dto.rfa.ApplicantDTO;
 import gov.ca.cwds.cals.service.dto.rfa.RFA1aFormDTO;
 import gov.ca.cwds.cals.service.dto.rfa.RFAAddressDTO;
+import gov.ca.cwds.cals.util.RfaAddressUtil;
 import gov.ca.cwds.cals.util.Utils;
 import gov.ca.cwds.data.legacy.cms.entity.PlacementHome;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
 import org.apache.commons.lang3.StringUtils;
 import org.mapstruct.AfterMapping;
 import org.mapstruct.Mapper;
@@ -12,14 +16,10 @@ import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.factory.Mappers;
 
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
 /**
  * @author CWDS CALS API Team
  */
-@Mapper(imports = {LocalDateTime.class, Utils.class, StringUtils.class})
+@Mapper(imports = {LocalDateTime.class, Utils.class, RfaAddressUtil.class, StringUtils.class})
 public interface PlacementHomeMapper {
   PlacementHomeMapper INSTANCE = Mappers.getMapper(PlacementHomeMapper.class);
 
@@ -27,15 +27,12 @@ public interface PlacementHomeMapper {
   @Mapping(target = "ageToNo", constant = "0")
   @Mapping(target = "atCapInd", constant = "N")
   @Mapping(target = "bckPersnm", constant = " ")
-  @Mapping(target = "bckExtNo", ignore = true)
-  @Mapping(target = "bckTelNo", ignore = true)
   @Mapping(target = "chlcrPlcd", constant = "U")
   @Mapping(target = "cityNm", source = "residentialAddress.city")
   @Mapping(target = "clSrvdc", constant = "0")
   @Mapping(target = "confEfind", constant = "N")
   @Mapping(target = "curOcpNo", constant = "0")
   @Mapping(target = "emrShltcd", constant = "U")
-  @Mapping(target = "faxNo", ignore = true)
   @Mapping(target = "frgAdrtB", constant = "N")
   @Mapping(target = "gndrAcpcd", constant = " ")
   @Mapping(target = "geoRgntcd", constant = " ")
@@ -49,12 +46,10 @@ public interface PlacementHomeMapper {
   @Mapping(target = "licBsnc", constant = "0")
   @Mapping(target = "licnseeNm", constant = " ")
   @Mapping(target = "licensrCd", constant = "CT")
-  @Mapping(target = "facltyNm", ignore = true)
   @Mapping(target = "pCityNm", constant = " ")
   @Mapping(target = "pyeFstnm", constant = " ")
   @Mapping(target = "pyeLstnm", constant = " ")
   @Mapping(target = "pyeMidnm", constant = " ")
-  @Mapping(target = "payeeStateCode", ignore = true)
   @Mapping(target = "pstreetNm", constant = " ")
   @Mapping(target = "pstreetNo", constant = " ")
   //@Mapping(target = "pZipNo", constant = "0")
@@ -67,8 +62,8 @@ public interface PlacementHomeMapper {
   @Mapping(target = "pvdTrnscd", constant = "U")
   @Mapping(target = "pubTrnscd", constant = "U")
   @Mapping(target = "stateCode", source = "residentialAddress.state.cwsId")
-  @Mapping(target = "streetNm", expression = "java(Utils.Address.getStreetName(residentialAddress))")
-  @Mapping(target = "streetNo", expression = "java(Utils.Address.getStreetNumber(residentialAddress))")
+  @Mapping(target = "streetNm", expression = "java(RfaAddressUtil.getStreetName(residentialAddress))")
+  @Mapping(target = "streetNo", expression = "java(RfaAddressUtil.getStreetNumber(residentialAddress))")
   @Mapping(target = "zipNo", source = "residentialAddress.zip")
   @Mapping(target = "addrDsc", source = "form.residence.directionsToHome")
   @Mapping(target = "spcharDsc", constant = " ")
@@ -93,22 +88,13 @@ public interface PlacementHomeMapper {
   //@Mapping(target = "laPZipno", constant = "0")
   //@Mapping(target = "laPZpsfx", constant = "0")
   @Mapping(target = "laPBsnss", constant = " ")
-  @Mapping(target = "laPPhNo", ignore = true)
-  @Mapping(target = "laPPhEx", ignore = true)
   @Mapping(target = "adhmonly", constant = " ")
-  @Mapping(target = "pyeExtNo", ignore = true)
-  @Mapping(target = "pyeTelNo", ignore = true)
   @Mapping(target = "arcassInd", constant = "N")
   @Mapping(target = "comfacInd", constant = "N")
   @Mapping(target = "trnhsgInd", constant = "N")
   @Mapping(target = "trnhsgFac", constant = "N")
-  @Mapping(target = "newlicNo", ignore = true)
   @Mapping(target = "newlicUpd", constant = "N")
-  @Mapping(target = "oldfacId", ignore = true)
   @Mapping(target = "emCntB", constant = "N")
-  @Mapping(target = "endDt", ignore = true)
-  @Mapping(target = "phEndc", ignore = true)
-  @Mapping(target = "endComdsc", ignore = true)
   PlacementHome toPlacementHome(RFA1aFormDTO form, RFAAddressDTO residentialAddress);
 
   @AfterMapping
