+++ src/main/java/gov/ca/cwds/cals/service/mapper/RFA1aFormMapper.java
@@ -41,18 +41,6 @@ public interface RFA1aFormMapper {
   @Mapping(target = "otherTypeDescription", source = "application.otherTypeDescription")
   @Mapping(target = "applicationCounty", source = "application.applicationCounty")
   @Mapping(target = "metadata", source = "application.metadata")
-  @Mapping(target = "applicantsDeclaration", ignore = true)
-  @Mapping(target = "references", ignore = true)
-  @Mapping(target = "applicantsHistory", ignore = true)
-  @Mapping(target = "adoptionHistory", ignore = true)
-  @Mapping(target = "applicants", ignore = true)
-  @Mapping(target = "minorChildren", ignore = true)
-  @Mapping(target = "otherAdults", ignore = true)
-  @Mapping(target = "rfa1cForms", ignore = true)
-  @Mapping(target = "applicantsRelationship", ignore = true)
-  @Mapping(target = "residence", ignore = true)
-  @Mapping(target = "childDesired", ignore = true)
-  @Mapping(target = "messages", ignore = true)
   RFA1aFormDTO toRFA1aFormDTO(RFA1aForm form);
 
   @Named("toExpandedRFA1aFormDTO")
@@ -80,23 +68,10 @@ public interface RFA1aFormMapper {
   @Mapping(target = "application.otherTypeDescription", source = "otherTypeDescription")
   @Mapping(target = "application.applicationCounty", source = "applicationCounty")
   @Mapping(target = "application.metadata", source = "metadata")
-  @Mapping(target = "application", ignore = true)
-  @Mapping(target = "applicantsDeclaration", ignore = true)
-  @Mapping(target = "references", ignore = true)
-  @Mapping(target = "applicantsHistory", ignore = true)
-  @Mapping(target = "adoptionHistory", ignore = true)
   @Mapping(target = "applicants", ignore = true)
   @Mapping(target = "minorChildren", ignore = true)
   @Mapping(target = "otherAdults", ignore = true)
   @Mapping(target = "rfa1cForms", ignore = true)
-  @Mapping(target = "applicantsRelationship", ignore = true)
-  @Mapping(target = "residence", ignore = true)
-  @Mapping(target = "childDesired", ignore = true)
-  @Mapping(target = "updateDateTime", ignore = true)
-  @Mapping(target = "updateUserId", ignore = true)
-  @Mapping(target = "createDateTime", ignore = true)
-  @Mapping(target = "createUserId", ignore = true)
-  @Mapping(target = "status", ignore = true)
   void toRFA1aForm(@MappingTarget RFA1aForm rfa1aForm, RFA1aFormDTO formDTO);
 
   @AfterMapping
